move AutoGenTestModel file to  /test/model dir
